+++ extensions/shottr/src/ssap.tsx
@@ -1,7 +1,9 @@
+import { closeMainWindow } from "@raycast/api";
 import { withShottrCheck } from "./utils/checkInstall";
 import { execSync } from "child_process";
 
 export default withShottrCheck(async function () {
   const url = "shottr://grab/append";
+  closeMainWindow();
   execSync(`open -g ${url}`);
 });